eracai/requirements.txt
weaviate-client==3.24.1
fastapi==0.95.0
uvicorn==0.21.0
python-dotenv==0.21.0
click==8.1.3
pydantic==1.10.12
asyncio
aiohttp
click==8.1.3
openai==1.3.0
anthropic==0.3.0
google-generativeai
groq
pytest
pytest-asyncio